+++ src/compas_viewer/configurations/__init__.py
@@ -1,8 +1,13 @@
 from .config import Config  # noqa: F401
-from .viewer_config import ViewerConfig, ViewerConfigType  # noqa: F401
 from .controller_config import ControllerConfig, ControllerConfigType  # noqa: F401
 from .render_config import RenderConfig, RenderConfigType  # noqa: F401
 from .render_config import CameraConfig, CameraConfigType  # noqa: F401
 from .render_config import SelectorConfig, SelectorConfigType  # noqa: F401
 from .controller_config import ActionConfig, ActionConfigType, MouseConfig, MouseConfigType  # noqa: F401
 from .scene_config import SceneConfig, SceneConfigType  # noqa: F401
+from .layout_config import LayoutConfig, LayoutConfigType  # noqa: F401
+from .layout_config import StatusbarConfig, StatusbarConfigType  # noqa: F401
+from .layout_config import WindowConfig, WindowConfigType  # noqa: F401
+from .layout_config import MenubarConfig, MenubarConfigType  # noqa: F401
+from .layout_config import ViewportConfig, ViewportConfigType  # noqa: F401
+from .layout_config import ToolbarConfig, ToolbarConfigType  # noqa: F401