+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+* Added Actions: delete_selected, camera_info, selection_info.
+* Added basic buildup of `Layout`'s configurations. The .json template and the configuration classes.
+* `Layout` class basic buildup: `Layout`, `ViewportLayout`, `WindowLayout`.
+* Added terminal activation of the viewer.
 * Added `PyOpenGL_accelerate` dependency.
 * Added repo images. 
 * Added `installation` documentation.
@@ -49,6 +53,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `NetworkObject`.
 
 ### Changed
+* `Timer` is moved to utilities.
+* Updated `BRepObject` to connect the latest OCC package, close [#48](https://github.com/compas-dev/compas_viewer/issues/48).
+* Fixed opacity bug in shaded mode.
 * Fixed main page link.
 * Fixed issue [#17](https://github.com/compas-dev/compas_viewer/issues/17) and avoid using `vertex_xyz`.
 * Update the dependency of `compas`.